backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.4
python-dotenv==1.0.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.9
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.2
email-validator==2.1.0
requests==2.32.3
httpx==0.27.0
aiofiles==23.2.1
libvirt-python==11.3.0
